+++ takamaka_memory/src/takamaka/memory/MemoryBlockchainTest.java
@@ -50,100 +50,104 @@ public class MemoryBlockchainTest {
 		TransactionReference takamaka_base = run("", "takamaka_base.jar", () -> blockchain.addJarStoreInitialTransaction(Paths.get("../takamaka_base/dist/takamaka_base.jar")));
 		Classpath takamakaBaseClasspath = new Classpath(takamaka_base, false);  // true/false irrelevant here
 		StorageReference gamete = run("", "gamete", () -> blockchain.addGameteCreationTransaction(takamakaBaseClasspath, BigInteger.valueOf(1000000)));
-		TransactionReference test_contracts_dependency = run("gamete", "test_contract_dependency.jar", () -> blockchain.addJarStoreTransaction(gamete, 10000, takamakaBaseClasspath, Paths.get("../test_contracts_dependency/dist/test_contracts_dependency.jar"), takamakaBaseClasspath));
-		TransactionReference test_contracts = run("gamete", "test_contracts.jar", () -> blockchain.addJarStoreTransaction(gamete, 10000, takamakaBaseClasspath, Paths.get("../test_contracts/dist/test_contracts.jar"), new Classpath(test_contracts_dependency, true))); // true relevant here
+		TransactionReference test_contracts_dependency = run("gamete", "test_contract_dependency.jar",
+			() -> blockchain.addJarStoreTransaction(gamete, BigInteger.valueOf(10000), takamakaBaseClasspath, Paths.get("../test_contracts_dependency/dist/test_contracts_dependency.jar"), takamakaBaseClasspath));
+		TransactionReference test_contracts = run("gamete", "test_contracts.jar",
+			() -> blockchain.addJarStoreTransaction(gamete, BigInteger.valueOf(10000), takamakaBaseClasspath, Paths.get("../test_contracts/dist/test_contracts.jar"), new Classpath(test_contracts_dependency, true))); // true relevant here
 		Classpath classpath = new Classpath(test_contracts, true);
 
-		StorageReference italianTime = run("gamete", "italianTime = new ItalianTime(13,25,40)", () -> blockchain.addConstructorCallTransaction(gamete, 20000, classpath, new ConstructorReference("takamaka.tests.ItalianTime", INT, INT, INT),
-				new IntValue(13), new IntValue(25), new IntValue(40)));
+		StorageReference italianTime = run("gamete", "italianTime = new ItalianTime(13,25,40)",
+			() -> blockchain.addConstructorCallTransaction(gamete, BigInteger.valueOf(20000), classpath, new ConstructorReference("takamaka.tests.ItalianTime", INT, INT, INT),
+			new IntValue(13), new IntValue(25), new IntValue(40)));
 		StorageReference wrapper1 = run("gamete", "wrapper1 = new Wrapper(italianTime)", () -> blockchain.addConstructorCallTransaction
-				(gamete, 20000, classpath,
+				(gamete, BigInteger.valueOf(20000), classpath,
 				new ConstructorReference("takamaka.tests.Wrapper", new ClassType("takamaka.tests.Time")),
 				italianTime));
 		run("gamete", "wrapper1.toString()", () -> (StringValue) blockchain.addInstanceMethodCallTransaction
-			(gamete, 20000, classpath,
+			(gamete, BigInteger.valueOf(20000), classpath,
 			new MethodReference(ClassType.OBJECT, "toString"),
 			wrapper1));
 		StorageReference wrapper2 = run("gamete", "wrapper2 = new Wrapper(italianTime,\"hello\",13011973,12345L)", () -> blockchain.addConstructorCallTransaction
-				(gamete, 20000, classpath,
+				(gamete, BigInteger.valueOf(20000), classpath,
 				new ConstructorReference("takamaka.tests.Wrapper", new ClassType("takamaka.tests.Time"), ClassType.STRING, ClassType.BIG_INTEGER, BasicTypes.LONG),
 				italianTime, new StringValue("hello"), new BigIntegerValue(BigInteger.valueOf(13011973)), new LongValue(12345L)));
 		run("gamete", "wrapper2.toString()", () -> (StringValue) blockchain.addInstanceMethodCallTransaction
-				(gamete, 20000, classpath,
+				(gamete, BigInteger.valueOf(20000), classpath,
 				new MethodReference(ClassType.OBJECT, "toString"),
 				wrapper2));
 		// we call the @Entry constructor Sub(int)
 		run("gamete", "new Sub(1973)", () -> blockchain.addConstructorCallTransaction
-			(gamete, 20000, classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
+			(gamete, BigInteger.valueOf(20000), classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
 
-		StorageReference sub1 = run("gamete", "sub1 = new Sub()", () -> blockchain.addConstructorCallTransaction(gamete, 100, classpath, new ConstructorReference("takamaka.tests.Sub")));
+		StorageReference sub1 = run("gamete", "sub1 = new Sub()",
+			() -> blockchain.addConstructorCallTransaction(gamete, BigInteger.valueOf(100), classpath, new ConstructorReference("takamaka.tests.Sub")));
 		// we try to call Sub.m1(): it is an entry that goes into a runtime exception
 		run("gamete", "sub1.m1()", () -> blockchain.addInstanceMethodCallTransaction
-				(gamete, 20000, classpath, new MethodReference("takamaka.tests.Sub", "m1"), sub1));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.tests.Sub", "m1"), sub1));
 
 		// we try to call a static method in the wrong way
 		run("gamete", "sub1.ms()", () -> blockchain.addInstanceMethodCallTransaction
-				(gamete, 20000, classpath, new MethodReference("takamaka.tests.Sub", "ms"), sub1));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.tests.Sub", "ms"), sub1));
 
 		// we call it in the correct way
 		run("gamete", "Sub.ms()", () -> blockchain.addStaticMethodCallTransaction
-			(gamete, 20000, classpath, new MethodReference("takamaka.tests.Sub", "ms")));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.tests.Sub", "ms")));
 
 		// we try to call an instance method as if it were static
 		run("gamete", "Sub.m5()", () -> blockchain.addStaticMethodCallTransaction
-				(gamete, 20000, classpath, new MethodReference("takamaka.tests.Sub", "m5")));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.tests.Sub", "m5")));
 
 		StorageReference eoa = run("gamete", "eoa = new ExternallyOwnedAccount()",
-				() -> blockchain.addConstructorCallTransaction(gamete, 20000, classpath, new ConstructorReference("takamaka.lang.ExternallyOwnedAccount")));
+			() -> blockchain.addConstructorCallTransaction(gamete, BigInteger.valueOf(20000), classpath, new ConstructorReference("takamaka.lang.ExternallyOwnedAccount")));
 
 		// we call the constructor Sub(int): it is @Entry but the caller has not enough funds
 		run("eoa", "new Sub(1973)", () -> blockchain.addConstructorCallTransaction
-				(eoa, 20000, classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
+			(eoa, BigInteger.valueOf(20000), classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
 
 		// we recharge eoa
 		run("gamete", "eoa.receive(2000)", () -> blockchain.addInstanceMethodCallTransaction
-				(gamete, 20000, classpath, new MethodReference("takamaka.lang.PayableContract", "receive", INT), eoa, new IntValue(2000)));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.lang.PayableContract", "receive", INT), eoa, new IntValue(2000)));
 
 		// still not enough funds since we are promising too much gas
 		run("eoa", "sub2 = new Sub(1973)", () -> blockchain.addConstructorCallTransaction
-			(eoa, 20000, classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
+			(eoa, BigInteger.valueOf(20000), classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
 
 		// now it's ok
 		StorageReference sub2 = run("eoa", "sub2 = new Sub(1973)", () -> blockchain.addConstructorCallTransaction
-			(eoa, 150, classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
+			(eoa, BigInteger.valueOf(150), classpath, new ConstructorReference("takamaka.tests.Sub", INT), new IntValue(1973)));
 
 		run("gamete", "sub2.print(italianTime)", () -> blockchain.addInstanceMethodCallTransaction
-			(gamete, 20000, classpath, new MethodReference("takamaka.tests.Sub", "print", new ClassType("takamaka.tests.Time")), sub2, italianTime));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.tests.Sub", "print", new ClassType("takamaka.tests.Time")), sub2, italianTime));
 		
 		run("gamete", "sub1.m4(13)", () -> blockchain.addInstanceMethodCallTransaction
-			(gamete, 20000, classpath, new MethodReference("takamaka.tests.Sub", "m4", INT), sub1, new IntValue(13)));
+			(gamete, BigInteger.valueOf(20000), classpath, new MethodReference("takamaka.tests.Sub", "m4", INT), sub1, new IntValue(13)));
 
 		// alias tests
 		ClassType alias = new ClassType("takamaka.tests.Alias");
-		StorageReference a1 = run("gamete", "a1 = new Alias()", () -> blockchain.addConstructorCallTransaction(gamete, 1000, classpath, new ConstructorReference(alias)));
-		StorageReference a2 = run("gamete", "a2 = new Alias()", () -> blockchain.addConstructorCallTransaction(gamete, 1000, classpath, new ConstructorReference(alias)));
+		StorageReference a1 = run("gamete", "a1 = new Alias()", () -> blockchain.addConstructorCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new ConstructorReference(alias)));
+		StorageReference a2 = run("gamete", "a2 = new Alias()", () -> blockchain.addConstructorCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new ConstructorReference(alias)));
 
 		// this test should return false
-		run("gamete", "a1.test(a1, a2)", () -> blockchain.addInstanceMethodCallTransaction(gamete, 1000, classpath, new MethodReference(alias, "test", alias, alias), a1, a1, a2));
+		run("gamete", "a1.test(a1, a2)", () -> blockchain.addInstanceMethodCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new MethodReference(alias, "test", alias, alias), a1, a1, a2));
 		// this test should return true
-		run("gamete", "a1.test(a1, a1)", () -> blockchain.addInstanceMethodCallTransaction(gamete, 1000, classpath, new MethodReference(alias, "test", alias, alias), a1, a1, a1));
+		run("gamete", "a1.test(a1, a1)", () -> blockchain.addInstanceMethodCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new MethodReference(alias, "test", alias, alias), a1, a1, a1));
 
 		StringValue s1 = new StringValue("hello");
 		StringValue s2 = new StringValue("hello");
 		ClassType string = new ClassType("java.lang.String");
 
 		// this test should return false
-		run("gamete", "a1.test(s1, s2)", () -> blockchain.addInstanceMethodCallTransaction(gamete, 1000, classpath, new MethodReference(alias, "test", string, string), a1, s1, s2));
+		run("gamete", "a1.test(s1, s2)", () -> blockchain.addInstanceMethodCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new MethodReference(alias, "test", string, string), a1, s1, s2));
 		// this test should return false since String parameters are considered different
-		run("gamete", "a1.test(s1, s1)", () -> blockchain.addInstanceMethodCallTransaction(gamete, 1000, classpath, new MethodReference(alias, "test", string, string), a1, s1, s1));		
+		run("gamete", "a1.test(s1, s1)", () -> blockchain.addInstanceMethodCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new MethodReference(alias, "test", string, string), a1, s1, s1));		
 
 		BigIntegerValue bi1 = new BigIntegerValue(BigInteger.valueOf(13));
 		BigIntegerValue bi2 = new BigIntegerValue(BigInteger.valueOf(13));
 		ClassType bigInteger = new ClassType("java.math.BigInteger");
 
 		// this test should return false
-		run("gamete", "a1.test(bi1, bi2)", () -> blockchain.addInstanceMethodCallTransaction(gamete, 1000, classpath, new MethodReference(alias, "test", bigInteger, bigInteger), a1, bi1, bi2));
+		run("gamete", "a1.test(bi1, bi2)", () -> blockchain.addInstanceMethodCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new MethodReference(alias, "test", bigInteger, bigInteger), a1, bi1, bi2));
 		// this test should return false since BigInteger parameters are considered different
-		run("gamete", "a1.test(bi1, bi1)", () -> blockchain.addInstanceMethodCallTransaction(gamete, 1000, classpath, new MethodReference(alias, "test", bigInteger, bigInteger), a1, bi1, bi1));		
+		run("gamete", "a1.test(bi1, bi1)", () -> blockchain.addInstanceMethodCallTransaction(gamete, BigInteger.valueOf(1000), classpath, new MethodReference(alias, "test", bigInteger, bigInteger), a1, bi1, bi1));		
 	}
 }
