+++ takamaka_runtime/src/takamaka/blockchain/OutOfGasException.java
@@ -1,7 +1,9 @@
 package takamaka.blockchain;
 
+import java.math.BigInteger;
+
 public class OutOfGasException extends RuntimeException {
-	public OutOfGasException(long amount) {
+	public OutOfGasException(BigInteger amount) {
 		super("Out of gas exception: " + amount);
 	}
 }
